requirements.txt
streamlit
dotenv
langchain_core
langchain
langchain_groq
pytesseract==0.3.10
pillow